LayoutTests/editing/deleting/delete-3608430-fix.html

<!DOCTYPE html>
<html> 
<head>

<style>
.editing { 
    border: 2px solid red; 
    padding: 12px; 
    font-size: 24px; 
}
</style>
<script src=../editing.js language="JavaScript" type="text/JavaScript" ></script>

<script>

function editingTest() {
    for (i = 0; i < 28; i++) {
        moveSelectionForwardByCharacterCommand();    
    }
    for (i = 0; i < 28; i++) {
        deleteCommand(); 
    }
}

</script>

<title>Editing Test</title> 
</head> 
<body>
<p>This test makes sure that deletion of an element auto-removes a blank element before it. Below, you should see "start-deleting", followed by several lines showing the states of the content-editable div over time, ending with 'start-deleting' inside a span id'd as "test".</p>
<div contenteditable id="root" class="editing">
<span id="test">foo 
<br>foo <i>  </i> <img src="../resources/abe.png"> select-in-me-and-start-deleting</span>
</div>

<script>
runDumpAsTextEditingTest();
</script>

</body>
</html>
