LayoutTests/platform/ios-simulator/editing/deleting/delete-3608430-fix-expected.txt

This test makes sure that deletion of an element auto-removes a blank element before it. Below, you should see "start-deleting", followed by several lines showing the states of the content-editable div over time, ending with 'start-deleting' inside a span id'd as "test".

foo 
foo   select-in-me-andstart-deleting
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
execDeleteCommand: <span id="test">foo <br>foo <i> </i> <img src="../resources/abe.png"> select-in-me-andstart-deleting</span>
